backEnd/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
alembic==1.13.1

flask-bcrypt==1.0.1

requests==2.32.3

flask-mail
